if文任意修正

+++ app/controllers/items_controller.rb
@@ -37,12 +37,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if item.user_id == current_user.id
-      item.destroy
-      redirect_to root_path
-    else
-      redirect_to root_path
-    end
+    item.destroy if item.user_id == current_user.id
+    redirect_to root_path
   end
 
   private
